+++ mallocx.c
@@ -52,7 +52,7 @@ GC_API int GC_CALL GC_get_kind_and_size(const void * p, size_t * psize)
     hdr * hhdr = HDR(p);
 
     if (psize != NULL) {
-        *psize = hhdr -> hb_sz;
+        *psize = (size_t)hhdr->hb_sz;
     }
     return hhdr -> hb_obj_kind;
 }
@@ -112,7 +112,7 @@ GC_API void * GC_CALL GC_realloc(void * p, size_t lb)
     }
     h = HBLKPTR(p);
     hhdr = HDR(h);
-    sz = hhdr -> hb_sz;
+    sz = (size_t)hhdr->hb_sz;
     obj_kind = hhdr -> hb_obj_kind;
     orig_sz = sz;
